all: remove redundant type conversion
GitHub-Last-Rev: 43904f8dd8f08028af8870fb8de4c1662594887e
GitHub-Pull-Request: golang/arch#6
Reviewed-on: https://go-review.googlesource.com/c/arch/+/428983
Run-TryBot: Cherry Mui <cherryyz@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>

+++ ppc64/ppc64asm/field.go
@@ -67,7 +67,7 @@ func (bs *BitFields) Append(b BitField) {
 // the sequence of bitfields is reasonable.
 func (bs BitFields) parse(i [2]uint32) (u uint64, Bits uint8) {
 	for _, b := range bs {
-		u = (uint64(u) << b.Bits) | uint64(b.Parse(i))
+		u = (u << b.Bits) | uint64(b.Parse(i))
 		Bits += b.Bits
 	}
 	return u, Bits
